requirements.txt
paho-mqtt
shortuuid
pytest
